requirements.txt
aiohttp==3.13.4
black==26.3.1
pre-commit==4.2.0
ruff==0.9.4
exa-py==2.4.0
google-genai==1.71.0
openai==1.78.1
pandas==2.2.3
parallel-web==0.4.1
pydantic==2.11.4
pytest==8.3.4
pytest-asyncio==0.24.0
python-dotenv==1.0.1
retry==0.9.2
tavily-python==0.7.21
tqdm==4.67.1
youdotcom==2.3.0
